Rename component files
use files named by the component instead of
index.tsx, which makes searching through the code
base for components harder
