+++ src/components/App/index.tsx
@@ -3,7 +3,7 @@ import './App.scss'
 import { gdprHandler } from '../../utils/gdpr'
 import Router from './router'
 import { IonApp, IonHeader } from '@ionic/react'
-import SwipeableTemporaryDrawer from '../../components/SwipeableTemporaryDrawer'
+import SwipeableTemporaryDrawer from '../SwipeableTemporaryDrawer/SwipeableTemporaryDrawer'
 import { Navbar } from 'react-bootstrap'
 import useDetectMobileOrTablet from '../../hooks/useDetectMobileOrTablet'
 import styled from 'styled-components'
