+++ src/components/Form/PhoneField.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
 import PropTypes from 'prop-types'
-import Country from '../Country'
+import Country from '../Country/Country'
 import Field from './Field'
 import { useField } from 'formik'
 import { getLocale } from '../../i18n/helper'
