+++ src/layout/Columns/Columns.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import useDocumentTitle from '../../hooks/useDocumentTitle'
 import { SCREEN } from '../../styles/theme'
-import Footer from '../../components/Footer'
-import BaseLang from '../../components/Lang'
+import Footer from '../../components/Footer/Footer'
+import BaseLang from '../../components/Lang/Lang'
 
 const Lang = styled(BaseLang)`
     position: absolute;
@@ -27,7 +27,8 @@ const Left = styled.div`
     width: 100%;
     background: ${({ theme }) => theme.color.white};
     ${SCREEN.DESKTOP} {
-        padding: ${({ theme }) => `${theme.spacing.XXXL} ${theme.spacing.M} ${theme.spacing.XXXL} ${theme.spacing.M}`};
+        padding: ${({ theme }) =>
+            `${theme.spacing.XXXL} ${theme.spacing.M} ${theme.spacing.XXXL} ${theme.spacing.M}`};
         width: 55%;
     }
 `
