+++ src/pages/PersonalData/PersonalData.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { useTranslation } from 'react-i18next'
 import ReactMarkdown from 'react-remarkable'
 
-import DefaultLayout from '../../layout/Default'
+import DefaultLayout from '../../layout/Default/Default'
 import { IonContent } from '@ionic/react'
 
 const PersonalData = () => {
