+++ src/pages/VideoCall/VideoCall.tsx
@@ -4,9 +4,9 @@ import DailyIframe from '@daily-co/daily-js'
 import { useTranslation } from 'react-i18next'
 import dailyCssText from './dailyCssText'
 import { CallContainer } from './VideoCallComponents'
-import Fullscreen from '../../components/Fullscreen'
+import Fullscreen from '../../components/Fullscreen/Fullscreen'
 import sendCallLogs from '../../actions/sendCallLogs'
-import Footer from '../../components/Footer'
+import Footer from '../../components/Footer/Footer'
 import { hideSupport, setVideoCallExited } from '../../utils/support'
 import Controls from './Controls'
 import VideoCallFrame from './VideoCallFrame'
