+++ src/pages/index.tsx
@@ -1,12 +1,12 @@
-import Home from './Home'
-import VideoCall from './VideoCall'
-import LegalMentions from './LegalMentions'
-import PersonalData from './PersonalData'
-import Blog from './Blog'
-import Credits from './Credits'
-import MediaNews from './MediaNews'
-import NotFound from './NotFound'
-import JoinVideoCall from './JoinVideoCall'
+import Home from './Home/Home'
+import VideoCall from './VideoCall/VideoCall'
+import LegalMentions from './LegalMentions/LegalMentions'
+import PersonalData from './PersonalData/PersonalData'
+import Blog from './Blog/Blog'
+import Credits from './Credits/Credits'
+import MediaNews from './MediaNews/MediaNews'
+import NotFound from './NotFound/NotFound'
+import JoinVideoCall from './JoinVideoCall/JoinVideoCall'
 
 export {
     Home,
